You can pick or take photos, refactor for more demos in the future
